I finished scene 1 for the startup menu
I added one more option, I also changed the aspect ratios of the title,
menu. I renamed all of the objects as well. All the buttons on the start
menu work. One thing you guys might want to try is to build and run
black scene, it is way to zoomed in.